refactor: restructure postgresql module
- create transactions blueprint
- rename postgresql module to transactions
- implement controllers for transactions blueprint
- moved transactions related models to transactions/models.py
